feat: 审核管理&修复统计管理

+++ src/main/java/com/crm/entity/Approval.java
@@ -9,12 +9,13 @@ import com.baomidou.mybatisplus.annotation.TableName;
 import java.time.LocalDateTime;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
+import io.swagger.models.auth.In;
 import lombok.Getter;
 import lombok.Setter;
 
 /**
  * <p>
- * 
+ *
  * </p>
  *
  * @author crm
@@ -32,7 +33,7 @@ public class Approval {
 
     @ApiModelProperty("0-合同审核，1-回款审核")
     @TableField("type")
-    private Byte type;
+    private Integer type;
 
     @ApiModelProperty("创建人id")
     @TableField("creater_id")
@@ -57,7 +58,7 @@ public class Approval {
     @ApiModelProperty("逻辑删除 0-未删除，1-已删除")
     @TableField(value = "delete_flag", fill = FieldFill.INSERT)
     @TableLogic
-    private Byte deleteFlag;
+    private Integer deleteFlag;
 
     @ApiModelProperty("创建时间")
     @TableField(value = "create_time", fill = FieldFill.INSERT)
